Moves single-use function back to reducer file
